requirements.txt
requests==2.7.0
python-decouple==2.3
pytest==2.7.2
